update: 주식명, 주식 정보 설정

+++ src/main/java/com/capstone/withyou/controller/UserInfoController.java
@@ -26,7 +26,7 @@ public class UserInfoController {
 
     // 보유 금액 업데이트
     @PutMapping("/{userId}/balance")
-    public ResponseEntity<Void> updateBalance(@PathVariable String userId, @RequestBody BigDecimal balance) {
+    public ResponseEntity<Void> updateBalance(@PathVariable String userId, @RequestBody Double balance) {
         userInfoService.updateUserBalance(userId, balance);
         return ResponseEntity.noContent().build();
     }
